Fix the provider nav
