+++ cfme/tests/infrastructure/test_provisioning_dialog.py
@@ -3,7 +3,7 @@ import pytest
 import re
 from datetime import datetime, timedelta
 
-from cfme.infrastructure.provisioning import provisioning_form
+from cfme.web_ui.prov_form import provisioning_form
 from cfme.infrastructure.virtual_machines import Vm, details_page
 from cfme.services import requests
 from cfme.web_ui import fill, flash
