+++ cfme/tests/infrastructure/test_pxe_provisioning.py
@@ -1,7 +1,7 @@
 import pytest
 
 from utils.conf import cfme_data
-from cfme.infrastructure.provisioning import provisioning_form
+from cfme.web_ui.prov_form import provisioning_form
 from cfme.infrastructure.pxe import get_pxe_server_from_config, get_template_from_config
 from cfme.services import requests
 from cfme.web_ui import flash, fill
