+++ cfme/tests/services/test_direct_links.py
@@ -4,7 +4,7 @@ import pytest
 import random
 
 from cfme.fixtures import pytest_selenium as sel
-from cfme.infrastructure.provisioning import provisioning_form
+from cfme.web_ui.prov_form import provisioning_form
 from cfme.services import requests
 from cfme.web_ui import flash
 from utils.browser import browser
